mandatory part

+++ ft_memccpy.c
@@ -6,15 +6,15 @@
 /*   By: egoncalv <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/23 02:57:11 by egoncalv          #+#    #+#             */
-/*   Updated: 2021/11/20 11:47:36 by egoncalv         ###   ########.fr       */
+/*   Updated: 2021/11/20 12:02:23 by egoncalv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-//#include "libft.h"
+#include "libft.h"
 
-void	*ft_memccpy(void *dst, const void *src, int c, size_t n)
+void	*ft_memccpy(void *dst, const void *src, int c, size_t  n)
 {
-	int		i;
+	size_t		i;
 
 	i = 0;
 	while (i < n && dst && src)
