+++ ft_memmove.c
@@ -6,15 +6,15 @@
 /*   By: egoncalv <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/23 02:34:39 by egoncalv          #+#    #+#             */
-/*   Updated: 2021/11/20 11:47:59 by egoncalv         ###   ########.fr       */
+/*   Updated: 2021/11/20 12:04:06 by egoncalv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-//#include "libft.h"
+#include "libft.h"
 
 void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < len)
